refactor(spec): split mapPathGroup into union and keyword modules
Extract anyOf/const union classification and keyword handlers from the
TypeBox agent error mapper so fall-through form keys stay explicit via a
done/continue result. Public mapValueErrors diagnostics are unchanged.
