+++ .changeset/spec-validate-map-path-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split TypeBox path-group error mapping into union classification and keyword handlers. Public validate() agent diagnostics are unchanged.
